feat(rbac): add audit log for RBAC backend (#1726)
* feat(rbac): add audit log for RBAC backend


* feat(rbac): simplify code


* feat(rbac): handle code review feeback


* fix(rbac): remove extra audit log for condition-storage


* feat(rbac): make audit log for all endpoints


* feat(rbac): simplify code


* feat(rbac): clean up code


* feat(rbac): fix unit tests, clean up code


* feat(rbac): fix code


* feat(rbac): improve code


* feat(rbac): fix tests


* feat(rbac): fix sonar cloud issue


* feat(rbac): improve audit log messages


* feat(rbac): handle more unit tests to check audit log


* feat(rbac): fix unit tests after rebase


* feat(rbac): use released version audit-log lib


* feat(rbac): add small audit log doc


* Update packages/backend/src/logger/customLogger.ts


* Update packages/backend/src/logger/customLogger.ts


* Update packages/backend/src/logger/customLogger.ts


* Update packages/backend/src/logger/customLogger.ts


* Update packages/backend/src/logger/customLogger.ts


* Update plugins/rbac-backend/src/file-permissions/csv-file-watcher.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/file-permissions/csv-file-watcher.ts


* Update plugins/rbac-backend/src/file-permissions/csv-file-watcher.ts


* Update plugins/rbac-backend/src/file-permissions/csv-file-watcher.ts


* Update plugins/rbac-backend/src/file-permissions/csv-file-watcher.ts


* Update plugins/rbac-backend/src/helper.ts


* Update plugins/rbac-backend/src/helper.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* Update plugins/rbac-backend/src/service/permission-policy.ts


* feat(rbac): fix grammar for doc


* fix(rbac): handle code review feeback


---------

+++ plugins/rbac-backend/src/service/policies-rest-api.ts
@@ -25,11 +25,12 @@ import {
   MetadataResponse,
 } from '@backstage/plugin-permission-node';
 
-import { Router } from 'express';
+import express from 'express';
 import { Request } from 'express-serve-static-core';
 import { isEmpty, isEqual } from 'lodash';
 import { ParsedQs } from 'qs';
 
+import { AuditLogger } from '@janus-idp/backstage-plugin-audit-log-node';
 import {
   PermissionAction,
   PermissionInfo,
@@ -45,6 +46,15 @@ import {
 } from '@janus-idp/backstage-plugin-rbac-common';
 import { PluginIdProvider } from '@janus-idp/backstage-plugin-rbac-node';
 
+import {
+  ConditionEvents,
+  ListConditionEvents,
+  ListPluginPoliciesEvents,
+  PermissionEvents,
+  RoleEvents,
+  SEND_RESPONSE_STAGE,
+} from '../audit-log/audit-logger';
+import { auditError as logAuditError } from '../audit-log/rest-errors-interceptor';
 import { ConditionalStorage } from '../database/conditional-storage';
 import {
   daoToMetadata,
@@ -73,6 +83,7 @@ export class PoliciesServer {
     private readonly conditionalStorage: ConditionalStorage,
     private readonly pluginIdProvider: PluginIdProvider,
     private readonly roleMetadata: RoleMetadataStorage,
+    private readonly aLog: AuditLogger,
   ) {}
 
   private async authorize(
@@ -81,9 +92,9 @@ export class PoliciesServer {
     permission: ResourcePermission,
   ): Promise<PolicyDecision> {
     if (permission !== policyEntityReadPermission) {
-      const user = await identity.getIdentity({ request });
-      if (!user) {
-        throw new NotAllowedError('User not found');
+      const userIdentity = await identity.getIdentity({ request });
+      if (!userIdentity) {
+        throw new NotAllowedError('User identity not found');
       }
     }
 
@@ -99,7 +110,7 @@ export class PoliciesServer {
     return decision;
   }
 
-  async serve(): Promise<Router> {
+  async serve(): Promise<express.Router> {
     const router = await createRouter(this.options);
 
     const { identity, discovery, logger, config } = this.options;
@@ -168,7 +179,18 @@ export class PoliciesServer {
         policies = await this.enforcer.getPolicy();
       }
 
-      response.json(await this.transformPolicyArray(...policies));
+      const body = await this.transformPolicyArray(...policies);
+
+      await this.aLog.auditLog({
+        message: `Return list permission policies`,
+        eventName: PermissionEvents.GET_POLICY,
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 200, body },
+      });
+
+      response.json(body);
     });
 
     router.get(
@@ -188,7 +210,18 @@ export class PoliciesServer {
 
         const policy = await this.enforcer.getFilteredPolicy(0, entityRef);
         if (policy.length !== 0) {
-          response.json(await this.transformPolicyArray(...policy));
+          const body = await this.transformPolicyArray(...policy);
+
+          await this.aLog.auditLog({
+            message: `Return permission policy`,
+            eventName: PermissionEvents.GET_POLICY,
+            stage: SEND_RESPONSE_STAGE,
+            status: 'succeeded',
+            request,
+            response: { status: 200, body },
+          });
+
+          response.json(body);
         } else {
           throw new NotFoundError(); // 404
         }
@@ -222,6 +255,17 @@ export class PoliciesServer {
         const processedPolicies = await this.processPolicies(policyRaw, true);
 
         await this.enforcer.removePolicies(processedPolicies);
+
+        await this.aLog.auditLog({
+          message: `Deleted permission policies`,
+          eventName: PermissionEvents.DELETE_POLICY,
+          metadata: { policies: processedPolicies, source: 'rest' },
+          stage: SEND_RESPONSE_STAGE,
+          status: 'succeeded',
+          request,
+          response: { status: 204 },
+        });
+
         response.status(204).end();
       },
     );
@@ -245,7 +289,23 @@ export class PoliciesServer {
 
       const processedPolicies = await this.processPolicies(policyRaw);
 
+      const entityRef = processedPolicies[0][0];
+      const roleMetadata = await this.roleMetadata.findRoleMetadata(entityRef);
+      if (entityRef.startsWith('role:default') && !roleMetadata) {
+        throw new Error(`Corresponding role ${entityRef} was not found`);
+      }
+
       await this.enforcer.addPolicies(processedPolicies, 'rest');
+
+      await this.aLog.auditLog({
+        message: `Created permission policies`,
+        eventName: PermissionEvents.CREATE_POLICY,
+        metadata: { policies: processedPolicies, source: 'rest' },
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 201 },
+      });
 
       response.status(201).end();
     });
@@ -313,11 +373,27 @@ export class PoliciesServer {
           'new policy',
         );
 
+        const roleMetadata =
+          await this.roleMetadata.findRoleMetadata(entityRef);
+        if (entityRef.startsWith('role:default') && !roleMetadata) {
+          throw new Error(`Corresponding role ${entityRef} was not found`);
+        }
+
         await this.enforcer.updatePolicies(
           processedOldPolicy,
           processedNewPolicy,
           'rest',
         );
+
+        await this.aLog.auditLog({
+          message: `Updated permission policies`,
+          eventName: PermissionEvents.UPDATE_POLICY,
+          metadata: { policies: processedNewPolicy, source: 'rest' },
+          stage: SEND_RESPONSE_STAGE,
+          status: 'succeeded',
+          request,
+          response: { status: 200 },
+        });
 
         response.status(200).end();
       },
@@ -338,7 +414,18 @@ export class PoliciesServer {
 
       const roles = await this.enforcer.getGroupingPolicy();
 
-      response.json(await this.transformRoleArray(...roles));
+      const body = await this.transformRoleArray(...roles);
+
+      await this.aLog.auditLog({
+        message: `Return list roles`,
+        eventName: RoleEvents.GET_ROLE,
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 200, body },
+      });
+
+      response.json(body);
     });
 
     router.get('/roles/:kind/:namespace/:name', async (request, response) => {
@@ -359,7 +446,18 @@ export class PoliciesServer {
       );
 
       if (role.length !== 0) {
-        response.json(await this.transformRoleArray(...role));
+        const body = await this.transformRoleArray(...role);
+
+        await this.aLog.auditLog({
+          message: `Return ${body[0].name}`,
+          eventName: RoleEvents.GET_ROLE,
+          stage: SEND_RESPONSE_STAGE,
+          status: 'succeeded',
+          request,
+          response: { status: 200, body },
+        });
+
+        response.json(body);
       } else {
         throw new NotFoundError(); // 404
       }
@@ -414,14 +512,29 @@ export class PoliciesServer {
       }
 
       const user = await identity.getIdentity({ request });
+      const modifiedBy = user?.identity.userEntityRef!;
       const metadata: RoleMetadataDao = {
         roleEntityRef: roleRaw.name,
         source: 'rest',
         description: roleRaw.metadata?.description ?? '',
-        author: user?.identity.userEntityRef,
-        modifiedBy: user?.identity.userEntityRef,
+        author: modifiedBy,
+        modifiedBy,
       };
+
       await this.enforcer.addGroupingPolicies(roles, metadata);
+
+      await this.aLog.auditLog({
+        message: `Created ${metadata.roleEntityRef}`,
+        eventName: RoleEvents.CREATE_ROLE,
+        metadata: {
+          ...metadata,
+          members: roles.map(gp => gp[0]),
+        },
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 201 },
+      });
 
       response.status(201).end();
     });
@@ -472,7 +585,7 @@ export class PoliciesServer {
         ...newRoleRaw.metadata,
         source: newRoleRaw.metadata?.source ?? 'rest',
         roleEntityRef: newRoleRaw.name,
-        modifiedBy: user?.identity.userEntityRef,
+        modifiedBy: user?.identity.userEntityRef!,
       };
 
       const oldMetadata =
@@ -546,6 +659,23 @@ export class PoliciesServer {
 
       await this.enforcer.updateGroupingPolicies(oldRole, newRole, newMetadata);
 
+      let message = `Updated ${oldMetadata.roleEntityRef}.`;
+      if (newMetadata.roleEntityRef !== oldMetadata.roleEntityRef) {
+        message = `${message}. Role entity reference renamed to ${newMetadata.roleEntityRef}`;
+      }
+      await this.aLog.auditLog({
+        message,
+        eventName: RoleEvents.UPDATE_ROLE,
+        metadata: {
+          ...newMetadata,
+          members: newRole.map(gp => gp[0]),
+        },
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 200 },
+      });
+
       response.status(200).end();
     });
 
@@ -583,20 +713,38 @@ export class PoliciesServer {
           }
         }
 
-        const metadata =
+        const currentMetadata =
           await this.roleMetadata.findRoleMetadata(roleEntityRef);
-
-        const err = await validateSource('rest', metadata);
+        const err = await validateSource('rest', currentMetadata);
         if (err) {
           throw new NotAllowedError(`Unable to delete role: ${err.message}`);
         }
 
         const user = await identity.getIdentity({ request });
+        const metadata: RoleMetadataDao = {
+          roleEntityRef,
+          source: 'rest',
+          modifiedBy: user?.identity.userEntityRef!,
+        };
+
         await this.enforcer.removeGroupingPolicies(
           roleMembers,
-          user?.identity.userEntityRef,
+          metadata,
           false,
         );
+
+        await this.aLog.auditLog({
+          message: `Deleted ${metadata.roleEntityRef}`,
+          eventName: RoleEvents.DELETE_ROLE,
+          metadata: {
+            ...metadata,
+            members: roleMembers.map(gp => gp[0]),
+          },
+          stage: SEND_RESPONSE_STAGE,
+          status: 'succeeded',
+          request,
+          response: { status: 204 },
+        });
 
         response.status(204).end();
       },
@@ -613,8 +761,18 @@ export class PoliciesServer {
         throw new NotAllowedError(); // 403
       }
 
-      const policies = await pluginPermMetaData.getPluginPolicies(this.auth);
-      response.json(policies);
+      const body = await pluginPermMetaData.getPluginPolicies(this.auth);
+
+      await this.aLog.auditLog({
+        message: `Return list plugin policies`,
+        eventName: ListPluginPoliciesEvents.GET_PLUGINS_POLICIES,
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 200, body },
+      });
+
+      response.json(body);
     });
 
     router.get('/plugins/condition-rules', async (request, response) => {
@@ -628,8 +786,18 @@ export class PoliciesServer {
         throw new NotAllowedError(); // 403
       }
 
-      const rules = await pluginPermMetaData.getPluginConditionRules(this.auth);
-      response.json(rules);
+      const body = await pluginPermMetaData.getPluginConditionRules(this.auth);
+
+      await this.aLog.auditLog({
+        message: `Return list conditional rules and schemas`,
+        eventName: ListConditionEvents.GET_CONDITION_RULES,
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 200, body },
+      });
+
+      response.json(body);
     });
 
     router.get('/roles/conditions', async (request, response) => {
@@ -650,14 +818,24 @@ export class PoliciesServer {
         this.getActionQueries(request.query.actions),
       );
 
-      const result: RoleConditionalPolicyDecision<PermissionAction>[] =
+      const body: RoleConditionalPolicyDecision<PermissionAction>[] =
         conditions.map(condition => {
           return {
             ...condition,
             permissionMapping: condition.permissionMapping.map(pm => pm.action),
           };
         });
-      response.json(result);
+
+      await this.aLog.auditLog({
+        message: `Return list conditional permission policies`,
+        eventName: ConditionEvents.GET_CONDITION,
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 200, body },
+      });
+
+      response.json(body);
     });
 
     router.post('/roles/conditions', async (request, response) => {
@@ -683,7 +861,19 @@ export class PoliciesServer {
       const id =
         await this.conditionalStorage.createCondition(conditionToCreate);
 
-      response.status(201).json({ id: id });
+      const body = { id: id };
+
+      await this.aLog.auditLog({
+        message: `Created conditional permission policy`,
+        eventName: ConditionEvents.CREATE_CONDITION,
+        metadata: { condition: conditionToCreate },
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 201, body },
+      });
+
+      response.status(201).json(body);
     });
 
     router.get('/roles/conditions/:id', async (request, response) => {
@@ -707,12 +897,21 @@ export class PoliciesServer {
         throw new NotFoundError();
       }
 
-      const result: RoleConditionalPolicyDecision<PermissionAction> = {
+      const body: RoleConditionalPolicyDecision<PermissionAction> = {
         ...condition,
         permissionMapping: condition.permissionMapping.map(pm => pm.action),
       };
 
-      response.json(result);
+      await this.aLog.auditLog({
+        message: `Return conditional permission policy by id`,
+        eventName: ConditionEvents.GET_CONDITION,
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 200, body },
+      });
+
+      response.json(body);
     });
 
     router.delete('/roles/conditions/:id', async (request, response) => {
@@ -731,7 +930,28 @@ export class PoliciesServer {
         throw new InputError('Id is not a valid number.');
       }
 
+      const condition = await this.conditionalStorage.getCondition(id);
+      if (!condition) {
+        throw new NotFoundError(`Condition with id ${id} was not found`);
+      }
+      const conditionToDelete: RoleConditionalPolicyDecision<PermissionAction> =
+        {
+          ...condition,
+          permissionMapping: condition.permissionMapping.map(pm => pm.action),
+        };
+
       await this.conditionalStorage.deleteCondition(id);
+
+      await this.aLog.auditLog({
+        message: `Deleted conditional permission policy`,
+        eventName: ConditionEvents.DELETE_CONDITION,
+        metadata: { condition: conditionToDelete },
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 204 },
+      });
+
       response.status(204).end();
     });
 
@@ -762,8 +982,21 @@ export class PoliciesServer {
       );
 
       await this.conditionalStorage.updateCondition(id, conditionToUpdate);
+
+      await this.aLog.auditLog({
+        message: `Updated conditional permission policy`,
+        eventName: ConditionEvents.UPDATE_CONDITION,
+        metadata: { conditionId: id, condition: conditionToUpdate },
+        stage: SEND_RESPONSE_STAGE,
+        status: 'succeeded',
+        request,
+        response: { status: 200 },
+      });
+
       response.status(200).end();
     });
+
+    router.use(logAuditError(this.aLog));
 
     return router;
   }
@@ -1014,11 +1247,6 @@ export class PoliciesServer {
           }' and action ${JSON.stringify(action)}`,
         );
       }
-      console.log(
-        `Found permission ${JSON.stringify(perm)} for resource type ${
-          roleConditionPolicy.resourceType
-        } and action ${action}`,
-      );
       permInfo.push({ name: perm.name, action });
     }
